night 22/07/2020 23h40

+++ src/lesson16_input_output_text_file/bai_tap/copy_file_text/CopyFileText.java
@@ -12,6 +12,7 @@ public class CopyFileText {
         String lickTarget = scanner.nextLine();
         CopyFileText copy = new CopyFileText();
         copy.copyFileText(linkSource, lickTarget);
+
     }
 
 
@@ -21,22 +22,29 @@ public class CopyFileText {
 
             File fileSource = new File(sourceFile);
             File fileTarget = new File(targetFile);
+            if(fileSource.exists() ){
+                if(!fileTarget.exists()) {
+                    fileTarget.createNewFile();
+                    BufferedReader brSource = new BufferedReader(new FileReader(fileSource));
+                    BufferedWriter brTarget = new BufferedWriter(new FileWriter(fileTarget));
 
-            BufferedReader brSource = new BufferedReader(new FileReader(fileSource));
-            BufferedWriter brTarget = new BufferedWriter(new FileWriter(fileTarget));
-
-            int sum=0;
-            String lineSource = "";
-            while ((lineSource = brSource.readLine()) != null) {
-                System.out.println(lineSource);
-                brTarget.write(lineSource);
-                brTarget.write("\n");
-                sum++;
+                    int sum = 0;
+                    String lineSource = "";
+                    while ((lineSource = brSource.readLine()) != null) {
+                        System.out.println(lineSource);
+                        brTarget.write(lineSource);
+                        brTarget.write("\n");
+                        sum++;
+                    }
+                    System.out.println("so ky tu vua copy sang la :" + sum);
+                    brSource.close();
+                    brTarget.close();
+                }else {
+                    System.err.println("file target already exists");
+                }
+            } else {
+                System.err.println("file Source not search ");
             }
-            System.out.println("so ky tu vua copy sang la :"+sum);
-            brSource.close();
-            brTarget.close();
-
 
         } catch (IOException e) {
             System.err.println("Fie không tồn tại or nội dung có lỗi!");
